Moved impl Serialize for Card to card.rs, deleting serialization.rs
